feat:add disputes apis changes on review
